+++ src/module/handlebarsHelpers.ts
@@ -1,5 +1,6 @@
 import { SWADE } from './config';
 import SwadeItem from './entities/SwadeItem';
+import { ItemType } from './enums/ItemTypeEnum';
 
 export const registerCustomHelpers = function () {
   Handlebars.registerHelper('add', function (a, b) {
@@ -36,7 +37,9 @@ export const registerCustomHelpers = function () {
   });
 
   Handlebars.registerHelper('enrich', (content) => {
-    return new Handlebars.SafeString(TextEditor.enrichHTML(content, {}));
+    return new Handlebars.SafeString(
+      TextEditor.enrichHTML(content, { secrets: true }),
+    );
   });
 
   Handlebars.registerHelper('canBeEquipped', (item: SwadeItem) => {
@@ -45,6 +48,39 @@ export const registerCustomHelpers = function () {
 
   Handlebars.registerHelper('disabled', (value) => {
     return value ? 'disabled' : '';
+  });
+
+  Handlebars.registerHelper('displayEmbedded', (array: any[] = []) => {
+    const collection = new Map(array);
+    const entities: string[] = [];
+    collection.forEach((val: any, key: string) => {
+      const type =
+        val.type === ItemType.Ability
+          ? game.i18n.localize('SWADE.SpecialAbility')
+          : game.i18n.localize(`ITEM.Type${val.type.capitalize()}`);
+
+      let majorMinor = '';
+      if (val.type === ItemType.Hindrance) {
+        if (val.data.major) {
+          majorMinor = game.i18n.localize('SWADE.Major');
+        } else {
+          majorMinor = game.i18n.localize('SWADE.Minor');
+        }
+      }
+
+      entities.push(
+        `<li class="flexrow">
+          <img src="${val.img}" alt="${type}" class="effect-icon" />
+          <span class="effect-label">${type} - ${val.name} ${majorMinor}</span>
+          <span class="effect-controls">
+            <a class="effect-action delete-embedded" data-Id="${key}">
+              <i class="fas fa-trash"></i>
+            </a>
+          </span>
+        </li>`,
+      );
+    });
+    return `<ul class="effects-list">${entities.join('\n')}</ul>`;
   });
 
   Handlebars.registerHelper('capitalize', (str) => {