+++ src/module/sheets/SwadeItemSheet.ts
@@ -1,5 +1,7 @@
+import { SWADE } from '../config';
 import SwadeEntityTweaks from '../dialog/entity-tweaks';
 import SwadeItem from '../entities/SwadeItem';
+import { AbilitySubtype } from '../enums/AbilitySubtypeEnum';
 import { ItemType } from '../enums/ItemTypeEnum';
 
 /**
@@ -67,6 +69,14 @@ export default class SwadeItemSheet extends ItemSheet {
 
   activateListeners(html) {
     super.activateListeners(html);
+
+    if (!this.isEditable) return;
+    if (
+      this.item.type === ItemType.Ability &&
+      this.item.data.data.subtype === AbilitySubtype.Race
+    ) {
+      this.form.ondrop = (ev) => this._onDrop(ev);
+    }
 
     // Delete Item from within Sheet. Only really used for Skills, Edges, Hindrances and Powers
     html.find('.item-delete').on('click', (ev) => {
@@ -154,6 +164,16 @@ export default class SwadeItemSheet extends ItemSheet {
         );
       }
     });
+
+    html.find('.delete-embedded').on('click', (ev) => {
+      ev.preventDefault();
+      const id = ev.currentTarget.dataset.id;
+      const map = new Map(
+        this.item.getFlag('swade', 'embeddedAbilities') || [],
+      );
+      map.delete(id);
+      this.item.setFlag('swade', 'embeddedAbilities', Array.from(map));
+    });
   }
 
   /**
@@ -163,7 +183,7 @@ export default class SwadeItemSheet extends ItemSheet {
   getData() {
     const data: any = super.getData();
     data.data.isOwned = this.item.isOwned;
-    data.config = CONFIG.SWADE;
+    data.config = SWADE;
     const actor = this.item.actor;
     const ownerIsWildcard = actor && actor.isWildcard;
     if (ownerIsWildcard || !this.item.isOwned) {
@@ -174,7 +194,10 @@ export default class SwadeItemSheet extends ItemSheet {
       attr['isCheckbox'] = attr['dtype'] === 'Boolean';
     }
     data.hasAdditionalStatsFields = Object.keys(additionalStats).length > 0;
-    data.displayNav = this.item.type !== ItemType.Skill;
+    data.displayNav = ![
+      ItemType.Skill.toString(),
+      ItemType.Ability.toString(),
+    ].includes(this.item.type);
 
     // Check for enabled optional rules
     data['settingrules'] = {
@@ -189,5 +212,62 @@ export default class SwadeItemSheet extends ItemSheet {
         break;
     }
     return data;
+  }
+
+  /**
+   * @override
+   */
+  async _onDrop(event: DragEvent) {
+    event.preventDefault();
+    event.stopPropagation();
+    let data;
+    let item: SwadeItem;
+
+    //get the data and accept it
+    try {
+      //get the data
+      data = JSON.parse(event.dataTransfer.getData('text/plain'));
+      if ('pack' in data) {
+        const pack = game.packs.get(data.pack) as Compendium;
+        item = (await pack.getEntity(data.id)) as SwadeItem;
+      } else if ('actorId' in data) {
+        item = new SwadeItem(data.data, {});
+      } else {
+        item = game.items.get(data.id) as SwadeItem;
+      }
+
+      const itemIsRightType = [
+        ItemType.Ability.toString(),
+        ItemType.Hindrance.toString(),
+        ItemType.Edge.toString(),
+        ItemType.Skill.toString(),
+      ].includes(item.type);
+
+      if (
+        data.type !== 'Item' ||
+        !itemIsRightType ||
+        (item.type === ItemType.Ability &&
+          item.data.data.subtype === AbilitySubtype.Race)
+      ) {
+        console.log(
+          'SWADE | Races only accept abilities, hindrances, edges and skills',
+        );
+        return false;
+      }
+    } catch (error) {
+      console.error(error);
+      return false;
+    }
+    //prep item data
+    const itemData = duplicate(item.data);
+    delete itemData['_id'];
+    delete itemData['permission'];
+
+    //pull the array from the flags, and push the new entry into it
+    const collection = this.item.getFlag('swade', 'embeddedAbilities') || [];
+    collection.push([randomID(), itemData]);
+    //save array back into flag
+    await this.item.setFlag('swade', 'embeddedAbilities', collection);
+    return false;
   }
 }